CW-397 Move Chatwoot Website Token to secrets
