+++ lib/src/screens/support/support_page.dart
@@ -6,6 +6,7 @@ import 'package:cake_wallet/utils/device_info.dart';
 import 'package:cake_wallet/utils/support_urls.dart';
 import 'package:flutter/material.dart';
 import 'package:url_launcher/url_launcher.dart';
+import 'package:cake_wallet/.secrets.g.dart' as secrets;
 
 class SupportPage extends BasePage {
   final imageLiveSupport = Image.asset('assets/images/live_support.png');
@@ -35,7 +36,7 @@ class SupportPage extends BasePage {
                   onPressed: () {
                     if (DeviceInfo.instance.isDesktop) {
                       _launchUrl(
-                          "$CHATWOOT_WEBSITE_TOKEN/widget?website_token=$CHATWOOT_WEBSITE_TOKEN");
+                          "$CHATWOOT_BASE_URL/widget?website_token=${secrets.chatwootWebsiteToken}");
                     } else {
                       Navigator.pushNamed(context, Routes.supportLiveChat);
                     }
