+++ lib/src/screens/support_chat/support_chat_page.dart
@@ -1,10 +1,11 @@
 import 'package:cake_wallet/generated/i18n.dart';
 import 'package:cake_wallet/src/screens/base_page.dart';
+import 'package:cake_wallet/src/screens/support_chat/widgets/chatwoot_widget.dart';
 import 'package:cake_wallet/utils/support_urls.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
+import 'package:cake_wallet/.secrets.g.dart' as secrets;
 
-import 'chatwoot_widget.dart';
 
 class SupportChatPage extends BasePage {
   SupportChatPage({required this.secureStorage});
@@ -26,7 +27,7 @@ class SupportChatPage extends BasePage {
             return ChatwootWidget(
               secureStorage,
               baseUrl: CHATWOOT_BASE_URL,
-              websiteToken: CHATWOOT_WEBSITE_TOKEN,
+              websiteToken: secrets.chatwootWebsiteToken,
               authToken: snapshot.data!,
             );
           return Container();
