+++ lib/utils/support_urls.dart
@@ -1,3 +1,2 @@
 const SUPPORT_GUIDES_URL = "https://guides.cakewallet.com/";
 const CHATWOOT_BASE_URL = "https://support.cakewallet.com";
-const CHATWOOT_WEBSITE_TOKEN = "kyUCS2fX2xnPSRrtTixPyJWu";
