+++ tool/utils/secret_key.dart
@@ -31,6 +31,7 @@ class SecretKey {
     SecretKey('anonPayReferralCode', () => ''),
     SecretKey('fiatApiKey', () => ''),
     SecretKey('payfuraApiKey', () => ''),
+    SecretKey('chatwootWebsiteToken', () => ''),
   ];
 
   static final ethereumSecrets = [
